requirements.txt
streamlit
openai
numpy
openai
num2words
matplotlib
plotly
scipy
scikit-learn
pandas
tiktoken
azure-search-documents
python-dotenv
azure-identity
cohere
azure-ai-vision-imageanalysis
azure-storage-blob